requirements.txt
torch
torchvision
pandas
pillow
matplotlib
numpy
flask
flask-cors